Fix a lot of SLAM server Actually start NT4 client Add tests to SLAM Still need to fix oscillation

+++ src/main/java/com/team2383/SLAM/server/vision/VisionIONorthstar.java
@@ -48,10 +48,10 @@ public class VisionIONorthstar implements VisionIO {
 
     public void updateInputs(VisionIOInputs inputs, Pose3d robotPose) {
         var queue = observationSubscriber.readQueue();
-        inputs.timestamps = new long[queue.length];
+        inputs.timestamps = new double[queue.length];
         inputs.frames = new double[queue.length][];
         for (int i = 0; i < queue.length; i++) {
-            inputs.timestamps[i] = queue[i].timestamp / 1000000;
+            inputs.timestamps[i] = queue[i].timestamp / 1000000.0;
             inputs.frames[i] = queue[i].value;
         }
         inputs.fps = fpsSubscriber.get();
